quant-research-pipeline/requirements-extract.txt
marker-pdf>=1.10,<2.0
google-genai>=1.0,<2.0
pypdf>=4.0,<5.0
pyyaml>=6.0,<7.0
pydantic>=2.0,<3.0
python-dotenv>=1.0,<2.0
typer>=0.12,<1.0
rich>=13.0,<14.0